Split main package into several files
